Cutscene subtitle positioning fix for resolutions other than 1024x768.  Code by MikeStar.

+++ fs2_open/code/camera/camera.h
@@ -55,7 +55,7 @@ private:
 	void clone(const subtitle &sub);
 
 	std::vector<std::string> text_lines;
-	int text_pos[2];
+	float text_pos[2];
 
 	float display_time;
 	float fade_time;
